Cambiando parseInt por parseFloat

+++ 1-EntradaSalida/jsEntradaSalida-09.js
@@ -5,7 +5,7 @@ mostrar el importe con un aumento del 10 %
 en el cuadro de texto "RESULTADO".*/
 function mostrarAumento() {
 	var inputResult = document.getElementById("txtIdResultado");
-	var sueldo = parseInt(document.getElementById("txtIdSueldo").value);
+	var sueldo = parseFloat(document.getElementById("txtIdSueldo").value);
 	var aumento = (sueldo * 10) / 100;
 	inputResult.value = sueldo + aumento;
 }
