+++ 1-EntradaSalida/jsEntradaSalida-10.js
@@ -5,7 +5,7 @@ mostrar el importe con un Descuento del 25 %
 en el cuadro de texto "RESULTADO"*/
 function mostrarAumento() {
 	var resultInput = document.getElementById("txtIdResultado");
-	var importe = parseInt(document.getElementById("txtIdImporte").value);
+	var importe = parseFloat(document.getElementById("txtIdImporte").value);
 	var descuento = (importe * 25) / 100;
 	resultInput.value = importe - descuento;
 }
